\name{pkgnet-package}
\alias{pkgnet-package}
\docType{package}
\title{Network analysis of relations between R packages}
\description{
Package DESCRIPTION files allows for specifying several types of inter-package
relations.  These include fields like Depends, Suggests, Enhances etc. This
package allows for recovering graph structure based on these relations. Network
representation of R repositories enables the user to fruitfully explore the
interconnected space of available R functionality while the developers or
repository maintainers can quickly scan package forward and reverse
dependencies.
}
\author{
Author and maintainter: Michal Bojanowski \email{m.j.bojanowski@uu.nl}
}
\keyword{ package }
