% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgnet.R
\name{pkgnet}
\alias{pkgnet}
\alias{pkgnet.character}
\alias{pkgnet.default}
\alias{pkgnet.matrix}
\title{Build a network based on package availability matrix}
\usage{
pkgnet(object, ...)

\method{pkgnet}{default}(object, ...)

\method{pkgnet}{character}(object, ap_args = NULL, ...)

\method{pkgnet}{matrix}(object, enams = c("Depends", "Suggests", "Imports",
  "Enhances", "LinkingTo"), vnams = c("Version", "Priority", "License",
  "License_is_FOSS", "License_restricts_use", "OS_type", "Archs", "MD5sum",
  "NeedsCompilation", "File", "Repository"))
}
\arguments{
\item{object}{a matrix as returned by \code{\link{available.packages}} or a 
character scalar, one of \code{"cran"} or \code{"bioc"} to fetch and
process packages available on CRAN or on Bioconductor, or an URL to a
CRAN-like repository.}

\item{...}{arguments passed to/from other methods}

\item{ap_args}{\code{NULL} or list of arguments passed to
\code{\link{avaialble.packages}}}

\item{enams}{character, names of columns of \code{a} that are to be used as 
edge attributes}

\item{vnams}{character, names of columns of \code{a} that are to be used as 
vertex attributes}
}
\value{
Object of class \code{igraph}.
}
\description{
Given the matrix as returned by \code{available.packages} construct a graph,
of class \code{igraph} of inter-package relations.
}
\details{
The resulting graph (object of class \code{igraph}) is a multigraph: there
can be multiple relationships between any given pair of vertices. Different
types of relations can be disentagled using edge attribute called
\code{type}. It stores the type of relation as provided with \code{enams}
argument.
}
\examples{
\dontrun{
a <- available.packages(contrib.url("http://cran.r-project.org", "source"))
g <- pkgnet(a)
summary(g)
}


}
\seealso{
\code{\link{available.packages}}, \code{\link{graph.data.frame}}
}

