% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgnet-package.R
\docType{package}
\name{pkgnet-package}
\alias{pkgnet-package}
\title{Network based packages on CRAN}
\description{
This is a network object containing packages and relations between them. It
is based on a snapshot taken on February 18, 2010 around 18:30 GMT.
}
\details{
Network analysis of relations between R packages

Package DESCRIPTION files allows for specifying several types of
inter-package relations.  These include fields like Depends, Suggests,
Enhances etc. This package allows for recovering graph structure based on
these relations. Network representation of R repositories enables the user
to fruitfully explore the interconnected space of available R functionality
while the developers or repository maintainers can quickly scan package
forward and reverse dependencies.
}
\author{
Author and maintainter: Michal Bojanowski
}
\keyword{package}

