% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgnet-package.R
\docType{package}
\name{pkgnet-package}
\alias{pkgnet-package}
\title{Build and Analyze Network of R Packages}
\description{
Package DESCRIPTION files allows for specifying several types of 
inter-package relations. These include fields like Depends, Suggests, 
Enhances etc. This package and function \code{\link{pkgnet}} allows for 
recovering graph structure based on these relations. Network representation 
of R repositories enables the user to explore the interconnected space of 
available R functionality while the developers or repository maintainers can 
quickly scan package forward and reverse dependencies.
}
\author{
Author and maintainter: Michal Bojanowski
}
\keyword{package}

